took out the dark bar added logoutdialog to admin lobby and user lobby removed code from main activity
