+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
     android:label="@string/app_name"
     android:roundIcon="@mipmap/ic_launcher_round"
     android:supportsRtl="true"
-    android:theme="@style/Theme.AppCompat.DayNight.DarkActionBar"
+    android:theme="@style/Theme.LutraLegallyDistinctPocketMonsterArena"
     tools:targetApi="31">
     <activity
        android:name=".LoginActivity"
